rename /query to /tao, since we will put the query parsing and the api service together
