+++ src/tao/parser.rs
@@ -1,9 +1,9 @@
-use crate::query::ast::{Arg, AssocType, ObjType, Query, TaoArgs, TaoOp};
+use crate::tao::ast::{Arg, AssocType, ObjType, Query, TaoArgs, TaoOp};
 use pest::{self, Parser};
 use std::str::FromStr;
 
 #[derive(pest_derive::Parser)]
-#[grammar = "query/tao.pest"]
+#[grammar = "tao/tao.pest"]
 struct TaoParser;
 
 pub fn parse(source: &str) -> Vec<Query> {
